requirements.txt
fastapi>=0.98.0,<0.99.0
uvicorn>=0.22.0,<0.23.0
sqlmodel>=0.0.8,<0.0.9
pydantic>=1.10.8,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<0.0.7
alembic>=1.11.1,<1.12.0
pymysql>=1.1.0,<1.2.0
cryptography>=41.0.1,<42.0.0
email-validator>=2.0.0,<3.0.0
celery>=5.3.1,<5.4.0
redis>=4.6.0,<4.7.0
python-dotenv>=1.0.0,<1.1.0
gunicorn>=21.2.0,<21.3.0
httpx>=0.24.1,<0.25.0
pytest>=7.4.0,<7.5.0
pytest-cov>=4.1.0,<4.2.0
mypy>=1.4.1,<1.5.0
black>=23.7.0,<23.8.0
isort>=5.12.0,<5.13.0
flake8>=6.1.0,<6.2.0 